CpuUsage and MemUsage are now ints, representing a percentage used. The int is PERCENTAGE * 100, so it includes two decimal places of accuracy

+++ src/main/java/edu/depaul/armada/model/AgentContainerLog.java
@@ -22,8 +22,8 @@ public class AgentContainerLog implements Serializable {
 	public long memTotal;
 	public long diskUsed;
 	public long diskTotal;
-	public long cpuUsed;
-	public long cpuTotal;
+	public int cpuUsed;
+	public int cpuTotal;
 	public Timestamp timestamp;
 
 }
